test: rewrite tests in ESM TypeScript
