+++ .npmrc
@@ -1,2 +1,4 @@
 package-lock=false
 lockfile=true
+public-hoist-pattern[]=*@types/*
+public-hoist-pattern[]=axios
