+++ package.json
@@ -22,7 +22,7 @@
     "lint": "eslint ./ --ignore dist --fix",
     "prepare": "pnpm run build",
     "prepublishOnly": "pnpm run clean && pnpm run build",
-    "test": "node ./spec/test.cjs",
+    "test": "tsx ./spec/test.mts",
     "test.format": "prettier . --check",
     "test.lint": "eslint . --ignore-pattern spec/fixtures/**"
   },
@@ -62,17 +62,20 @@
     "@types/eslint-config-prettier": "^6.11.3",
     "@types/eslint__js": "^8.42.3",
     "@types/execa": "^2.0.0",
+    "@types/fs-extra": "~11.0.4",
     "@types/node": "^22.13.13",
     "@types/read-pkg-up": "^6.0.0",
     "@types/semver": "^7.7.0",
     "@types/shelljs": "^0.8.15",
     "eslint-config-atomic": "link:",
     "execa": "^5.1.1",
+    "fs-extra": "^11.3.0",
     "gitly": "^3.1.0",
     "npm-check-updates": "17.1.16",
     "prettier-config-atomic": "^4.0.0",
     "shelljs": "^0.9.2",
-    "shx": "^0.4.0"
+    "shx": "^0.4.0",
+    "tsx": "^4.19.3"
   },
   "peerDependencies": {
     "eslint": "^8 || ^7 || ^6"
